Use typed ConfigEntry in ring coordinator (#136457)
* Use typed ConfigEntry in ring coordinator

* Make config_entry a positional argument for coordinator
